feat: Add two geofence types
Can add landmark geofence within mall geofence.
