+++ app/src/main/java/com/example/triibe/triibeuserapp/trackLocation/AddMallGeofencesIntentService.java
@@ -1,0 +1,191 @@
+package com.example.triibe.triibeuserapp.trackLocation;
+
+import android.app.IntentService;
+import android.app.PendingIntent;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.util.Log;
+import android.widget.Toast;
+
+import com.example.triibe.triibeuserapp.util.Constants;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.common.api.ResultCallback;
+import com.google.android.gms.common.api.Status;
+import com.google.android.gms.location.Geofence;
+import com.google.android.gms.location.GeofencingRequest;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.maps.model.LatLng;
+
+import java.util.ArrayList;
+import java.util.Map;
+
+
+/**
+ * An {@link IntentService} subclass for handling asynchronous task requests in
+ * a service on a separate handler thread.
+ * <p/>
+ * TODO: Customize class - update intent actions and extra parameters.
+ */
+public class AddMallGeofencesIntentService extends IntentService
+        implements GoogleApiClient.ConnectionCallbacks,
+        GoogleApiClient.OnConnectionFailedListener,
+        ResultCallback<Status> {
+
+    private static final String TAG = "AddMallGeofences";
+
+    // TODO: Rename actions, choose action names that describe tasks that this
+    // IntentService can perform, e.g. ACTION_FETCH_NEW_ITEMS
+    public static final String ACTION_FOO = "com.example.triibe.triibeuserapp.trackLocation.action.FOO";
+    public static final String ACTION_BAZ = "com.example.triibe.triibeuserapp.trackLocation.action.BAZ";
+
+    // TODO: Rename parameters
+    public static final String EXTRA_PARAM1 = "com.example.triibe.triibeuserapp.trackLocation.extra.PARAM1";
+    public static final String EXTRA_PARAM2 = "com.example.triibe.triibeuserapp.trackLocation.extra.PARAM2";
+
+    GoogleApiClient mGoogleApiClient;
+    private ArrayList<Geofence> mGeofences;
+
+    public AddMallGeofencesIntentService() {
+        super("AddMallGeofencesIntentService");
+    }
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+        mGoogleApiClient = new GoogleApiClient.Builder(this)
+                .addConnectionCallbacks(this)
+                .addOnConnectionFailedListener(this)
+                .addApi(LocationServices.API)
+                .build();
+        mGeofences = new ArrayList<>();
+    }
+
+    @Override
+    protected void onHandleIntent(Intent intent) {
+
+        for (Map.Entry<String, LatLng> entry : Constants.TEST_MALLS.entrySet()) {
+            mGeofences.add(new Geofence.Builder()
+                    .setRequestId(entry.getKey())
+                    .setCircularRegion(
+                            entry.getValue().latitude,
+                            entry.getValue().longitude,
+                            Constants.GEOFENCE_MALL_RADIUS_IN_METERS
+                    )
+                    .setExpirationDuration(Constants.GEOFENCE_EXPIRATION_IN_MILLISECONDS)
+                    .setTransitionTypes(Geofence.GEOFENCE_TRANSITION_ENTER |
+                            Geofence.GEOFENCE_TRANSITION_EXIT)
+                    .build());
+        }
+
+        if (!mGoogleApiClient.isConnected()) {
+            mGoogleApiClient.connect();
+        } else {
+            addGeofences();
+        }
+
+//        if (intent != null) {
+//            final String action = intent.getAction();
+//            if (ACTION_FOO.equals(action)) {
+//                final String param1 = intent.getStringExtra(EXTRA_PARAM1);
+//                final String param2 = intent.getStringExtra(EXTRA_PARAM2);
+//                handleActionFoo(param1, param2);
+//            } else if (ACTION_BAZ.equals(action)) {
+//                final String param1 = intent.getStringExtra(EXTRA_PARAM1);
+//                final String param2 = intent.getStringExtra(EXTRA_PARAM2);
+//                handleActionBaz(param1, param2);
+//            }
+//        }
+    }
+
+    /**
+     * Handle action Foo in the provided background thread with the provided
+     * parameters.
+     */
+    private void handleActionFoo(String param1, String param2) {
+        // TODO: Handle action Foo
+        throw new UnsupportedOperationException("Not yet implemented");
+    }
+
+    /**
+     * Handle action Baz in the provided background thread with the provided
+     * parameters.
+     */
+    private void handleActionBaz(String param1, String param2) {
+        // TODO: Handle action Baz
+        throw new UnsupportedOperationException("Not yet implemented");
+    }
+
+    private void addGeofences() {
+        if (!mGoogleApiClient.isConnected()) {
+            Toast.makeText(this, "Google API client not connected", Toast.LENGTH_SHORT).show();
+            return;
+        }
+
+        try {
+            LocationServices.GeofencingApi.addGeofences(
+                    mGoogleApiClient,
+                    getGeofencingRequest(),
+                    getGeofencePendingIntent()
+            ).setResultCallback(this);
+        } catch (SecurityException securityException) {
+            // Catch exception generated if the app does not use ACCESS_FINE_LOCATION permission.
+        }
+
+        // Mark geofences as added.
+        SharedPreferences preferences = getSharedPreferences(Constants.MALL_GEOFENCES, 0);
+        SharedPreferences.Editor editor = preferences.edit();
+        editor.putBoolean(Constants.MALL_GEOFENCES_ADDED, true);
+        editor.apply();
+    }
+
+    private GeofencingRequest getGeofencingRequest() {
+        GeofencingRequest.Builder builder = new GeofencingRequest.Builder();
+        builder.setInitialTrigger(GeofencingRequest.INITIAL_TRIGGER_ENTER);
+        builder.addGeofences(mGeofences);
+        return builder.build();
+    }
+
+    private PendingIntent getGeofencePendingIntent() {
+        Intent intent = new Intent(this, MallGeofenceTransitionIntentService.class);
+        return PendingIntent.getService(this, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+    }
+
+    @Override
+    public void onConnected(@Nullable Bundle bundle) {
+        addGeofences();
+    }
+
+    @Override
+    public void onConnectionSuspended(int i) {
+        Log.d(TAG, "onConnectionSuspended: googleApiClient");
+    }
+
+    @Override
+    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+        Log.d(TAG, "onConnectionFailed: could not connect");
+    }
+
+    @Override
+    public void onResult(@NonNull Status status) {
+        if (status.isSuccess()) {
+            Log.d(TAG, "onResult: successfully added or removed mall geofence");
+        } else {
+            Log.d(TAG, "onResult: error adding or removing mall geofence");
+        }
+    }
+
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+
+        if (mGoogleApiClient.isConnected()) {
+            mGoogleApiClient.disconnect();
+        }
+        Toast.makeText(this, "add mall geofences service done", Toast.LENGTH_SHORT).show();
+    }
+}
